+++ lib/pub_api_client.dart
@@ -1,5 +1,7 @@
 export 'package:oauth2/oauth2.dart' show Credentials;
 
+export 'src/constants.dart';
+export 'src/helpers/exceptions.dart';
 export 'src/helpers/google_packages.dart';
 export 'src/models/barrel.dart';
 export 'src/pub_api_client_base.dart';