+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 2.0.4
+
+### Added
+
+- Better exception messaging
+
+## 2.0.3
+
+### Added
+
+- Support for retrieve pub credentials on windows
+
 ## 2.0.2
 
 ### Added